backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
anthropic==0.40.0
firebase-admin==6.6.0
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.10.3
httpx==0.28.1
